+++ components/browser/icons/src/main/java/mozilla/components/browser/icons/IconRequest.kt
@@ -127,11 +127,6 @@ data class IconRequest(
             MICROSOFT_TILE,
 
             /**
-             * An icon found in Mozilla's "tippy top" list.
-             */
-            TIPPY_TOP,
-
-            /**
              * An icon found in Mozilla's Merino Manifest list.
              *
              * https://merino.services.mozilla.com/api/v1/manifest