+++ components/browser/icons/src/main/java/mozilla/components/browser/icons/ext/Uri.kt
@@ -6,7 +6,8 @@ package mozilla.components.browser.icons.ext
 
 import android.net.Uri
 
-// Make sure domain added here have the corresponding image_url in icons-top200.json
+// Make sure domain added here have the corresponding image_url in the Merino manifest
+// (manifest/manifest.json).
 private val commonDomain = listOf("wikipedia.org")
 
 /**